web/requirements.txt
celery
Mezzanine==4.2.3
#mezzanine-api
django-audiofield
django-import-export
django-storages
boto3
django-celery
python-dateutil
s3fs
lxml
html5lib
django-debug-toolbar
django-filebrowser
djangorestframework
markdown
django-filter
django-tinymce
